requirements.txt
matplotlib
mplcursors
numpy
pexpect
Pillow
psutil
python-dotenv
requests
